+++ src/engine/mod.rs
@@ -4,6 +4,7 @@ mod field;
 mod function;
 mod literal;
 mod network;
+mod plan;
 mod symbol;
 mod tape;
 mod value;
@@ -11,6 +12,7 @@ mod value;
 pub use evaluation::Evaluation;
 pub use field::{Field, Gradients};
 pub use network::Network;
+pub use plan::Plan;
 pub use symbol::Symbol;
 pub use value::Value;